Restructured tests to make faster in parallel
- also lowered concurrency on insert + update/delete because browser is slow to paint cypress expect statements causing userbase-js to timeout
